+++ Modules/Intent.Modules.Eventing.MassTransit/release-notes.md
@@ -1,3 +1,7 @@
+### Version 6.2.8
+
+- Improvement: Changed the private class `MessageToSend` to `sealed` inline with best practices
+
 ### Version 6.2.7
 
 - Improvement: Small updated to module internal code